Move model helpers to shared database utils
