+++ cvat/apps/dataset_manager/bindings.py
@@ -36,7 +36,6 @@ from cvat.apps.engine import models
 from cvat.apps.engine.cache import MediaCache
 from cvat.apps.engine.lazy_list import LazyList
 from cvat.apps.engine.media_io.frame_provider import FrameOutputType, TaskFrameProvider
-from cvat.apps.engine.model_utils import add_prefetch_fields
 from cvat.apps.engine.models import (
     AttributeSpec,
     AttributeType,
@@ -53,6 +52,7 @@ from cvat.apps.engine.models import (
     Task,
 )
 from cvat.apps.engine.rq import ImportRQMeta
+from cvat.utils.django_database import add_prefetch_fields
 
 from ..engine.log import ServerLogManager
 from .annotation import AnnotationIR, AnnotationManager, TrackManager
